requirements.txt
# === OCR ===
paddleocr==2.7.3
paddlepaddle>=2.6.2
pytesseract==0.3.10

# === Imagen ===
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4

# === Backend ===
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# === Exportación ===
openpyxl==3.1.2
reportlab==4.1.0
pandas==2.2.2

# === Utilidades ===
python-dateutil==2.9.0
imagehash==4.3.1
pydantic==2.7.1
loguru==0.7.2

# === Tests ===
pytest==8.2.0
httpx==0.27.0
